--==entities_are_everywhere==-- Entity component system is going to prevail in my whole API. I wanna cnahge tha way of memory distribution, objects access and creation. In my memory directory I've changed NewT, DelT, NewTArr, DelTArr; These templated functions now are methods of  memory allocator class. Also, they and other functions which take unknown templated number of arguments - use move&& semantic for objects creation. Before I did that there were problems with copying and arguments. I simpli need set arguments as "Args&& ... Arguments" and pass them to another functions via std::forward<Args>(Arguments)... About entities. I am going to make generic class of entity to store id for every object(suc as data resource, graphics object, game entity...) and for usage of custom memory allocation. So, I am going to change DataRes to DataEntity and use EntSys for global access --==/entities_are_everywhere==--

+++ src_cpp/core/nwl_random.cpp
@@ -1,7 +1,7 @@
 #include <nwl_pch.hpp>
 #include "nwl_random.h"
 
-std::mt19937 NWL::Random::s_RandEngine(time(nullptr));
+std::mt19937 NWL::Random::s_RandEngine(static_cast<unsigned int>(time(nullptr)));
 
 namespace NW
 {
